# Faster Board::get_all_legal_moves