--- a/rust/Cargo.toml
+++ b/rust/Cargo.toml
@@ -14,6 +14,7 @@ crate-type = ["cdylib", "rlib"]
 default = ["console_error_panic_hook"]
 
 [dependencies]
+lazy_static = "1.4.0"
 wasm-bindgen = "0.2.63"
 js-sys = "0.3.44"