--- a/rust/src/board.rs
+++ b/rust/src/board.rs
@@ -254,17 +254,39 @@ impl Board {
 
         let mut n_reverses: [u8; 64] = [0; 64];
 
-        let pattern: [usize; 8] = [0, 8, 16, 24, 32, 40, 48, 56];
+        let patterns_8: [[usize; 8]; 18] = [
+            // horizontal
+            [0, 8, 16, 24, 32, 40, 48, 56],
+            [1, 9, 17, 25, 33, 41, 49, 57],
+            [2, 10, 18, 26, 34, 42, 50, 58],
+            [3, 11, 19, 27, 35, 43, 51, 59],
+            [4, 12, 20, 28, 36, 44, 52, 60],
+            [5, 13, 21, 29, 37, 45, 53, 61],
+            [6, 14, 22, 30, 38, 46, 54, 62],
+            [7, 15, 23, 31, 39, 47, 55, 63],
+            // vertical
+            [0, 1, 2, 3, 4, 5, 6, 7],
+            [8, 9, 10, 11, 12, 13, 14, 15],
+            [16, 17, 18, 19, 20, 21, 22, 23],
+            [24, 25, 26, 27, 28, 29, 30, 31],
+            [32, 33, 34, 35, 36, 37, 38, 39],
+            [40, 41, 42, 43, 44, 45, 46, 47],
+            [48, 49, 50, 51, 52, 53, 54, 55],
+            [56, 57, 58, 59, 60, 61, 62, 63],
+            // diagonal
+            [0, 9, 18, 27, 36, 45, 54, 63],
+            [7, 14, 21, 28, 35, 42, 49, 56],
+        ];
 
-        {
+        for pattern in patterns_8.iter() {
             let mut index: usize = 0;
             for i in 0..8 {
                 let cell = coded_board[pattern[i]] as usize;
                 index += cell * 3usize.pow(i as u32);
             }
-            let n_reverses_line = parse_reverse_index(N_REVERSES_8[index]);
+
             for i in 0..8 {
-                n_reverses[pattern[i]] += n_reverses_line[i];
+                n_reverses[pattern[i]] += N_REVERSES_8[8*index+i] as u8;
             }
         }
 
@@ -516,6 +538,29 @@ mod tests {
             );
             assert_eq!(board.is_empty(1), false);
             assert_eq!(board.is_empty(1 << 63), true);
+        }
+
+        #[test]
+        fn get_n_reverses() {
+            let board = Board::create_from_str(
+                "
+                - - - - - - - -
+                - - - - - - - -
+                - - o - - - - -
+                - - o o o - - -
+                - - - o x - - -
+                - - - - - - - -
+                - - - - - - - -
+                - - - - - - - -
+            ",
+            );
+            let actual = board.get_n_reverses(&Player::Second);
+            let expected = [
+                0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0,
+                0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                0, 0, 0, 0, 0, 0, 0, 0,
+            ];
+            assert_eq!(actual, expected);
         }
     }