--- a/rust/src/board_reverse.rs
+++ b/rust/src/board_reverse.rs
@@ -15,7 +15,8 @@ lazy_static! {
     pub static ref N_REVERSES_5: [u64; SIZE_N_REVERSES_5] = [0; SIZE_N_REVERSES_5];
     pub static ref N_REVERSES_6: [u64; SIZE_N_REVERSES_6] = [0; SIZE_N_REVERSES_6];
     pub static ref N_REVERSES_7: [u64; SIZE_N_REVERSES_7] = [0; SIZE_N_REVERSES_7];
-    pub static ref N_REVERSES_8: Box<[u64; 8 * SIZE_N_REVERSES_8]> = Box::new(get_size_n_reverses_8());
+    pub static ref N_REVERSES_8: Box<[u64; 8 * SIZE_N_REVERSES_8]> =
+        Box::new(get_size_n_reverses_8());
 }
 
 fn get_size_n_reverses_8() -> [u64; 8 * SIZE_N_REVERSES_8] {
@@ -51,5 +52,19 @@ mod tests {
             2 * 3usize.pow(6) + 3usize.pow(5) + 2 * 3usize.pow(4) + 2 * 3usize.pow(3) + 2 * 3 + 1;
         let expected = [0, 0, 3, 0, 0, 0, 0, 1];
         assert_eq!(actual[(8 * index)..(8 * index + 8)], expected);
+    }
+
+    #[test]
+    fn static_n_reverses_8() {
+        use crate::board_reverse::N_REVERSES_8;
+        // o x - x x o x -
+        let index: usize =
+            2 * 3usize.pow(6) + 3usize.pow(5) + 2 * 3usize.pow(4) + 2 * 3usize.pow(3) + 2 * 3 + 1;
+        for i in (8 * index)..(8 * index + 8) {
+            println!("N_REVERSES_8[{}] = {}", i, N_REVERSES_8[i]);
+        }
+
+        let expected = [0, 0, 3, 0, 0, 0, 0, 1];
+        assert_eq!(N_REVERSES_8[(8 * index)..(8 * index + 8)], expected);
     }
 }