--- a/rust/src/lib.rs
+++ b/rust/src/lib.rs
@@ -1,8 +1,12 @@
 #![feature(test)]
 extern crate wasm_bindgen;
 
+#[macro_use]
+extern crate lazy_static;
+
 pub mod bitboard;
 pub mod board;
+mod board_reverse;
 mod game;
 pub mod game_tree;
 pub mod ggf;