```
test board::tests::benches::get_all_legal_moves           ... bench:       1,009 ns/iter (+/- 108)
test game_tree::tests::benches::alpha_beta_pruning_search ... bench:  25,493,334 ns/iter (+/- 1,056,217)
test game_tree::tests::benches::get_children              ... bench:       1,079 ns/iter (+/- 44)
```

```
test board::tests::benches::get_all_legal_moves           ... bench:         375 ns/iter (+/- 21)
test game_tree::tests::benches::alpha_beta_pruning_search ... bench:  15,463,689 ns/iter (+/- 708,250)
test game_tree::tests::benches::get_children              ... bench:         476 ns/iter (+/- 10)
```